Add general discipline, improve VoNC
the general discipline section was added to allow for discipline of members without removing them from the club or position. The vote of no confidence section was improved to allow the defense to refute the claims and made the process longer to allow more thought and discussion before the decision was made.